Update QoS AQM thresholds to support relative values or bytes
* (M) openconfig-qos-mem-mgmt.yang
  - Add 2 leaves representing percent values for RED min-threshold and max-threshold

This PR is to address two issues:
1. Since the first implementations of RED, there has been no consistent guidance on
the units to use for queue-length (and therefore minth and maxth).  Over the years
implementations have used temporal, data, or percent values.  This is to support percent
values, which are 'unit-less' and a good approach for an abstraction.
2. Percent values carry the benefit of reducing configuration: a minth or maxth specified
in bytes must be uniquely configured for every interface speed on a given platform.  When
specified in bytes, a single set of thresholds can be applied to any interface speed.